Started the maze project

+++ mario/game_command.cpp
@@ -1,6 +1,6 @@
 #include "game_command.h"
 
-game_command game_command::create_command(game_command_header header, std::vector<int> arguments)
+game_command game_command::create_command(game_command_header header, std::vector<uint64_t> arguments)
 {
 	game_command result;
